test(pricing-fetcher): in source
